Add accounts created using AddAccountDialog to ResourceStore
It wasn't possible to actually add new storage accounts to the
application - no action was connected to AddAccountDialog's 'Add
account' button.

Add an onSave prop to AddAccountDialog; it's invoked when the 'Add
account' button is clicked and the data in the dialog is valid. Add a
new action, addStorageAccount, that is invoked by ResourcePane when
onSave is called. Finally add a handler for the addStorageAccount action
to ResourceStore. With these modifications accounts created using
AddAccountDialog are added to ResourceStore.