Add a Gulp task generating Electron config files
Electron config files (package.json and init.js) had to be manually
created and copied to the dist directory. As a result, the app wasn't
runnable from a freshly-cloned repository out of the box.

Add electron-init.js to the repository. Add a Gulp task copying that
file to the dist directory, simplifying the project's package.json and
outputting it as Electron's package.json. The app can now be run from a
freshly-cloned repository out of the box.