Build: copy the Glyphicon font to the dist directory
The Glyphicon font had to be manually copied to the dist directory. As a
result, the app didn't look quite right when started from a
freshly-cloned repository.

Add a Gulp task copying the Glyphicon Halflings font (WOFF2 variant) to
the dist directory. The app now looks as expected when run from a
freshly-cloned repository.