Prevent flicker when switching tabs
When a tab was clicked, the click event first of all triggered the
switchToTab action, then bubbled up, causing the tab link's default event
handler to be invoked. Since all tab links had their href attribute set to
'#', the start page was loaded as a result. A split second later it was
replaced with the tab contents, leading to a short but noticeable flicker.

Call preventDefault on the click event when a tab is clicked. As a result,
the tab link's default event handler is never invoked, and the main pane
reloads only once (to show the clicked tab's contents).