Extract TabPane and TabContentsPane from TabList
TabList grew too complex for a single component: it was responsible for
responding to tab state updates, displaying a list of tabs, and a pane
with contents of the active tab.

Move some of TabList's logic into separate componets:
- TabContentsPane displays the contents of the active tab
- TabPane handles communication with TabStore, groups TabList and
  TabContentsPane, and passes them the data they need to render

As a result, TabList is now responsible only for displaying the list of
open tabs.